+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 ### Changed
+- Send notification count via Server-Sent Events (SSE)
 - Cache calls to k8s and graphite from the API's content endpoint
 - Cache permissions in Auth
 - Increased default worker threads in Auth and API